requirements.txt
stripe==2.23.0
jinja2
google-api-python-client==2.43.0
google-api-core==1.31.5
#google-cloud-storage==2.3.0 #not on gae python
google-cloud-storage==2.0.0

google-cloud-ndb==1.11.1
loguru==0.5.3

uvicorn==0.17.0
gunicorn==20.1.0
webapp2==3.0.0b1
loguru
